api/pyproject.toml
[project]
name = "localization-management-api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (~=0.115.12)",
    "pydantic>=2.11.5",
    "python-dotenv (~=1.1.0)",
    "supabase (~=2.15.2)",
    "uvicorn[standard] (~=0.34.3)",
]
